Generate all the models.
