+++ client/v2/common/models/catchpoint_abort_response.go
@@ -1,0 +1,7 @@
+package models
+
+// CatchpointAbortResponse
+type CatchpointAbortResponse struct {
+	// CatchupMessage catchup abort response string
+	CatchupMessage string `json:"catchup-message,omitempty"`
+}
