+++ client/v2/common/models/catchpoint_start_response.go
@@ -1,0 +1,7 @@
+package models
+
+// CatchpointStartResponse
+type CatchpointStartResponse struct {
+	// CatchupMessage catchup start response string
+	CatchupMessage string `json:"catchup-message,omitempty"`
+}
