+++ client/v2/common/models/error_response.go
@@ -1,0 +1,10 @@
+package models
+
+// ErrorResponse an error response with optional data field.
+type ErrorResponse struct {
+	// Data
+	Data *map[string]interface{} `json:"data,omitempty"`
+
+	// Message
+	Message string `json:"message,omitempty"`
+}
