+++ client/v2/common/models/post_transactions_response.go
@@ -1,0 +1,7 @@
+package models
+
+// PostTransactionsResponse transaction ID of the submission.
+type PostTransactionsResponse struct {
+	// Txid encoding of the transaction hash.
+	Txid string `json:"txId,omitempty"`
+}
