+++ client/v2/common/models/responsemodels.go
@@ -1,175 +1,10 @@
 package models
 
-// AccountParticipation describes the parameters used by this account in consensus protocol.
-type AccountParticipation struct {
-
-	// \[sel\] Selection public key (if any) currently registered for this round.
-	SelectionParticipationKey []byte `json:"selection-participation-key,omitempty"`
-
-	// \[voteFst\] First round for which this participation is valid.
-	VoteFirstValid uint64 `json:"vote-first-valid,omitempty"`
-
-	// \[voteKD\] Number of subkeys in each batch of participation keys.
-	VoteKeyDilution uint64 `json:"vote-key-dilution,omitempty"`
-
-	// \[voteLst\] Last round for which this participation is valid.
-	VoteLastValid uint64 `json:"vote-last-valid,omitempty"`
-
-	// \[vote\] root participation public key (if any) currently registered for this round.
-	VoteParticipationKey []byte `json:"vote-participation-key,omitempty"`
-}
-
-// AssetParams specifies the parameters for an asset.
-type AssetParams struct {
-
-	// \[c\] Address of account used to clawback holdings of this asset.  If empty, clawback is not permitted.
-	Clawback string `json:"clawback,omitempty"`
-
-	// The address that created this asset. This is the address where the parameters for this asset can be found, and also the address where unwanted asset units can be sent in the worst case.
-	Creator string `json:"creator"`
-
-	// \[dc\] The number of digits to use after the decimal point when displaying this asset. If 0, the asset is not divisible. If 1, the base unit of the asset is in tenths. If 2, the base unit of the asset is in hundredths, and so on. This value must be between 0 and 19 (inclusive).
-	Decimals uint64 `json:"decimals"`
-
-	// \[df\] Whether holdings of this asset are frozen by default.
-	DefaultFrozen bool `json:"default-frozen,omitempty"`
-
-	// \[f\] Address of account used to freeze holdings of this asset.  If empty, freezing is not permitted.
-	Freeze string `json:"freeze,omitempty"`
-
-	// \[m\] Address of account used to manage the keys of this asset and to destroy it.
-	Manager string `json:"manager,omitempty"`
-
-	// \[am\] A commitment to some unspecified asset metadata. The format of this metadata is up to the application.
-	MetadataHash []byte `json:"metadata-hash,omitempty"`
-
-	// \[an\] Name of this asset, as supplied by the creator.
-	Name string `json:"name,omitempty"`
-
-	// \[r\] Address of account holding reserve (non-minted) units of this asset.
-	Reserve string `json:"reserve,omitempty"`
-
-	// \[t\] The total number of units of this asset.
-	Total uint64 `json:"total"`
-
-	// \[un\] Name of a unit of this asset, as supplied by the creator.
-	UnitName string `json:"unit-name,omitempty"`
-
-	// \[au\] URL where more information about the asset can be retrieved.
-	Url string `json:"url,omitempty"`
-}
-
 type RawBlockJson struct {
 	Block string
 }
 type RawBlockMsgpack struct {
 	Block string `json:"url,omitempty"`
-}
-
-// Block defines model for Block.
-type Block struct {
-	Cert string `json:"cert"`
-
-	// \[gh\] hash to which this block belongs.
-	GenesisHash []byte `json:"genesis-hash"`
-
-	// \[gen\] ID to which this block belongs.
-	GenesisId string `json:"genesis-id"`
-
-	// Current block hash
-	Hash []byte `json:"hash"`
-
-	// Period on which the block was confirmed.
-	Period uint64 `json:"period"`
-
-	// \[prev\] Previous block hash.
-	PreviousBlockHash []byte `json:"previous-block-hash"`
-
-	// Address that proposed this block.
-	Proposer string `json:"proposer"`
-
-	// Fields relating to rewards,
-	Rewards BlockRewards `json:"rewards,omitempty"`
-
-	// \[rnd\] Current round on which this block was appended to the chain.
-	Round uint64 `json:"round"`
-
-	// \[seed\] Sortition seed.
-	Seed []byte `json:"seed"`
-
-	// \[ts\] Block creation timestamp in seconds since eposh
-	Timestamp uint64 `json:"timestamp"`
-
-	// \[txns\] list of transactions corresponding to a given round.
-	Transactions []Transaction `json:"transactions,omitempty"`
-
-	// \[txn\] TransactionsRoot authenticates the set of transactions appearing in the block. More specifically, it's the root of a merkle tree whose leaves are the block's Txids, in lexicographic order. For the empty block, it's 0. Note that the TxnRoot does not authenticate the signatures on the transactions, only the transactions themselves. Two blocks with the same transactions but in a different order and with different signatures will have the same TxnRoot.
-	TransactionsRoot []byte `json:"transactions-root"`
-
-	// \[tc\] TxnCounter counts the number of transactions committed in the ledger, from the time at which support for this feature was introduced.
-	//
-	// Specifically, TxnCounter is the number of the next transaction that will be committed after this block.  It is 0 when no transactions have ever been committed (since TxnCounter started being supported).
-	TxnCounter uint64 `json:"txn-counter,omitempty"`
-
-	// Fields relating to a protocol upgrade.
-	UpgradeState BlockUpgradeState `json:"upgrade-state,omitempty"`
-
-	// Fields relating to voting for a protocol upgrade.
-	UpgradeVote BlockUpgradeVote `json:"upgrade-vote,omitempty"`
-}
-
-// BlockRewards defines model for BlockRewards.
-type BlockRewards struct {
-
-	// \[fees\] accepts transaction fees, it can only spend to the incentive pool.
-	FeeSink string `json:"fee-sink"`
-
-	// \[rwcalr\] number of leftover MicroAlgos after the distribution of rewards-rate MicroAlgos for every reward unit in the next round.
-	RewardsCalculationRound uint64 `json:"rewards-calculation-round"`
-
-	// \[earn\] How many rewards, in MicroAlgos, have been distributed to each RewardUnit of MicroAlgos since genesis.
-	RewardsLevel uint64 `json:"rewards-level"`
-
-	// \[rwd\] accepts periodic injections from the fee-sink and continually redistributes them as rewards.
-	RewardsPool string `json:"rewards-pool"`
-
-	// \[rate\] Number of new MicroAlgos added to the participation stake from rewards at the next round.
-	RewardsRate uint64 `json:"rewards-rate"`
-
-	// \[frac\] Number of leftover MicroAlgos after the distribution of RewardsRate/rewardUnits MicroAlgos for every reward unit in the next round.
-	RewardsResidue uint64 `json:"rewards-residue"`
-}
-
-// BlockUpgradeState defines model for BlockUpgradeState.
-type BlockUpgradeState struct {
-
-	// \[proto\] The current protocol version.
-	CurrentProtocol string `json:"current-protocol"`
-
-	// \[nextproto\] The next proposed protocol version.
-	NextProtocol string `json:"next-protocol,omitempty"`
-
-	// \[nextyes\] Number of blocks which approved the protocol upgrade.
-	NextProtocolApprovals uint64 `json:"next-protocol-approvals,omitempty"`
-
-	// \[nextswitch\] Round on which the protocol upgrade will take effect.
-	NextProtocolSwitchOn uint64 `json:"next-protocol-switch-on,omitempty"`
-
-	// \[nextbefore\] Deadline round for this protocol upgrade (No votes will be consider after this round).
-	NextProtocolVoteBefore uint64 `json:"next-protocol-vote-before,omitempty"`
-}
-
-// BlockUpgradeVote defines model for BlockUpgradeVote.
-type BlockUpgradeVote struct {
-
-	// \[upgradeyes\] Indicates a yes vote for the current proposal.
-	UpgradeApprove bool `json:"upgrade-approve,omitempty"`
-
-	// \[upgradedelay\] Indicates the time between acceptance and execution.
-	UpgradeDelay uint64 `json:"upgrade-delay,omitempty"`
-
-	// \[upgradeprop\] Indicates a proposed upgrade.
-	UpgradePropose string `json:"upgrade-propose,omitempty"`
 }
 
 // Supply defines model for Supply.
@@ -183,215 +18,6 @@ type Supply struct {
 
 	// TotalMoney
 	TotalMoney uint64 `json:"total-money"`
-}
-
-// Transaction contains all fields common to all transactions and serves as an
-// envelope to all transactions type.
-// Definition:
-// data/transactions/signedtxn.go : SignedTxn
-// data/transactions/transaction.go : Transaction
-type Transaction struct {
-	// ApplicationTransaction fields for application transactions.
-	// Definition:
-	// data/transactions/application.go : ApplicationCallTxnFields
-	ApplicationTransaction TransactionApplication `json:"application-transaction,omitempty"`
-
-	// AssetConfigTransaction fields for asset allocation, re-configuration, and
-	// destruction.
-	// A zero value for asset-id indicates asset creation.
-	// A zero value for the params indicates asset destruction.
-	// Definition:
-	// data/transactions/asset.go : AssetConfigTxnFields
-	AssetConfigTransaction TransactionAssetConfig `json:"asset-config-transaction,omitempty"`
-
-	// AssetFreezeTransaction fields for an asset freeze transaction.
-	// Definition:
-	// data/transactions/asset.go : AssetFreezeTxnFields
-	AssetFreezeTransaction TransactionAssetFreeze `json:"asset-freeze-transaction,omitempty"`
-
-	// AssetTransferTransaction fields for an asset transfer transaction.
-	// Definition:
-	// data/transactions/asset.go : AssetTransferTxnFields
-	AssetTransferTransaction TransactionAssetTransfer `json:"asset-transfer-transaction,omitempty"`
-
-	// AuthAddr (sgnr) The address used to sign the transaction. This is used for
-	// rekeyed accounts to indicate that the sender address did not sign the
-	// transaction.
-	AuthAddr string `json:"auth-addr,omitempty"`
-
-	// CloseRewards (rc) rewards applied to close-remainder-to account.
-	CloseRewards uint64 `json:"close-rewards,omitempty"`
-
-	// ClosingAmount (ca) closing amount for transaction.
-	ClosingAmount uint64 `json:"closing-amount,omitempty"`
-
-	// ConfirmedRound round when the transaction was confirmed.
-	ConfirmedRound uint64 `json:"confirmed-round,omitempty"`
-
-	// CreatedApplicationIndex specifies an application index (ID) if an application
-	// was created with this transaction.
-	CreatedApplicationIndex uint64 `json:"created-application-index,omitempty"`
-
-	// CreatedAssetIndex specifies an asset index (ID) if an asset was created with
-	// this transaction.
-	CreatedAssetIndex uint64 `json:"created-asset-index,omitempty"`
-
-	// Fee (fee) Transaction fee.
-	Fee uint64 `json:"fee,omitempty"`
-
-	// FirstValid (fv) First valid round for this transaction.
-	FirstValid uint64 `json:"first-valid,omitempty"`
-
-	// GenesisHash (gh) Hash of genesis block.
-	GenesisHash []byte `json:"genesis-hash,omitempty"`
-
-	// GenesisId (gen) genesis block ID.
-	GenesisId string `json:"genesis-id,omitempty"`
-
-	// GlobalStateDelta (gd) Global state key/value changes for the application being
-	// executed by this transaction.
-	GlobalStateDelta []EvalDeltaKeyValue `json:"global-state-delta,omitempty"`
-
-	// Group (grp) Base64 encoded byte array of a sha512/256 digest. When present
-	// indicates that this transaction is part of a transaction group and the value is
-	// the sha512/256 hash of the transactions in that group.
-	Group []byte `json:"group,omitempty"`
-
-	// Id transaction ID
-	Id string `json:"id,omitempty"`
-
-	// IntraRoundOffset offset into the round where this transaction was confirmed.
-	IntraRoundOffset uint64 `json:"intra-round-offset,omitempty"`
-
-	// KeyregTransaction fields for a keyreg transaction.
-	// Definition:
-	// data/transactions/keyreg.go : KeyregTxnFields
-	KeyregTransaction TransactionKeyreg `json:"keyreg-transaction,omitempty"`
-
-	// LastValid (lv) Last valid round for this transaction.
-	LastValid uint64 `json:"last-valid"`
-
-	// Lease (lx) Base64 encoded 32-byte array. Lease enforces mutual exclusion of
-	// transactions. If this field is nonzero, then once the transaction is confirmed,
-	// it acquires the lease identified by the (Sender, Lease) pair of the transaction
-	// until the LastValid round passes. While this transaction possesses the lease, no
-	// other transaction specifying this lease can be confirmed.
-	Lease []byte `json:"lease,omitempty"`
-
-	// LocalStateDelta (ld) Local state key/value changes for the application being
-	// executed by this transaction.
-	LocalStateDelta []AccountStateDelta `json:"local-state-delta,omitempty"`
-
-	// Note (note) Free form data.
-	Note []byte `json:"note,omitempty"`
-
-	// PaymentTransaction fields for a payment transaction.
-	// Definition:
-	// data/transactions/payment.go : PaymentTxnFields
-	PaymentTransaction TransactionPayment `json:"payment-transaction,omitempty"`
-
-	// ReceiverRewards (rr) rewards applied to receiver account.
-	ReceiverRewards uint64 `json:"receiver-rewards,omitempty"`
-
-	// RekeyTo (rekey) when included in a valid transaction, the accounts auth addr
-	// will be updated with this value and future signatures must be signed with the
-	// key represented by this address.
-	RekeyTo string `json:"rekey-to,omitempty"`
-
-	// RoundTime time when the block this transaction is in was confirmed.
-	RoundTime uint64 `json:"round-time,omitempty"`
-
-	// Sender (snd) Sender's address.
-	Sender string `json:"sender"`
-
-	// SenderRewards (rs) rewards applied to sender account.
-	SenderRewards uint64 `json:"sender-rewards,omitempty"`
-
-	// Signature validation signature associated with some data. Only one of the
-	// signatures should be provided.
-	Signature TransactionSignature `json:"signature"`
-
-	// TxType (type) Indicates what type of transaction this is. Different types have
-	// different fields.
-	// Valid types, and where their fields are stored:
-	// * (pay) payment-transaction
-	// * (keyreg) keyreg-transaction
-	// * (acfg) asset-config-transaction
-	// * (axfer) asset-transfer-transaction
-	// * (afrz) asset-freeze-transaction
-	// * (appl) application-transaction
-	Type string `json:"tx-type"`
-}
-
-// TransactionAssetConfig defines model for TransactionAssetConfig.
-type TransactionAssetConfig struct {
-
-	// \[xaid\] ID of the asset being configured or empty if creating.
-	AssetId uint64 `json:"asset-id,omitempty"`
-
-	// AssetParams specifies the parameters for an asset.
-	//
-	// \[apar\] when part of an AssetConfig transaction.
-	//
-	// Definition:
-	// data/transactions/asset.go : AssetParams
-	Params AssetParams `json:"params,omitempty"`
-}
-
-// TransactionAssetFreeze defines model for TransactionAssetFreeze.
-type TransactionAssetFreeze struct {
-
-	// \[fadd\] Address of the account whose asset is being frozen or thawed.
-	Address string `json:"address"`
-
-	// \[faid\] ID of the asset being frozen or thawed.
-	AssetId uint64 `json:"asset-id"`
-
-	// \[afrz\] The new freeze status.
-	NewFreezeStatus bool `json:"new-freeze-status"`
-}
-
-// TransactionAssetTransfer defines model for TransactionAssetTransfer.
-type TransactionAssetTransfer struct {
-
-	// \[aamt\] Amount of asset to transfer. A zero amount transferred to self allocates that asset in the account's Assets map.
-	Amount uint64 `json:"amount"`
-
-	// \[xaid\] ID of the asset being transferred.
-	AssetId uint64 `json:"asset-id"`
-
-	// \[aclose\] Indicates that the asset should be removed from the account's Assets map, and specifies where the remaining asset holdings should be transferred.  It's always valid to transfer remaining asset holdings to the creator account.
-	CloseTo string `json:"close-to,omitempty"`
-
-	// \[arcv\] Recipient address of the transfer.
-	Receiver string `json:"receiver"`
-
-	// \[asnd\] The effective sender during a clawback transactions. If this is not a zero value, the real transaction sender must be the Clawback address from the AssetParams.
-	Sender string `json:"sender,omitempty"`
-
-	CloseAmount uint64 `json:"close-amount,omitempty"`
-}
-
-// TransactionKeyreg defines model for TransactionKeyreg.
-type TransactionKeyreg struct {
-
-	// \[nonpart\] Mark the account as participating or non-participating.
-	NonParticipation bool `json:"non-participation,omitempty"`
-
-	// \[selkey\] Public key used with the Verified Random Function (VRF) result during committee selection.
-	SelectionParticipationKey []byte `json:"selection-participation-key,omitempty"`
-
-	// \[votefst\] First round this participation key is valid.
-	VoteFirstValid uint64 `json:"vote-first-valid,omitempty"`
-
-	// \[votekd\] Number of subkeys in each batch of participation keys.
-	VoteKeyDilution uint64 `json:"vote-key-dilution,omitempty"`
-
-	// \[votelst\] Last round this participation key is valid.
-	VoteLastValid uint64 `json:"vote-last-valid,omitempty"`
-
-	// \[votekey\] Participation public key used in key registration transactions.
-	VoteParticipationKey []byte `json:"vote-participation-key,omitempty"`
 }
 
 // TransactionParams contains the parameters that help a client construct a new transaction.
@@ -419,93 +45,6 @@ type TransactionParams struct {
 	// The minimum transaction fee (not per byte) required for the
 	// txn to validate for the current network protocol.
 	MinFee uint64 `json:"min-fee,omitempty"`
-}
-
-// TransactionPayment defines model for TransactionPayment.
-type TransactionPayment struct {
-
-	// \[amt\] number of MicroAlgos intended to be transferred.
-	Amount uint64 `json:"amount"`
-
-	// Number of MicroAlgos that were sent to the close-remainder-to address when closing the sender account.
-	CloseAmount uint64 `json:"close-amount,omitempty"`
-
-	// \[close\] when set, indicates that the sending account should be closed and all remaining funds be transferred to this address.
-	CloseRemainderTo string `json:"close-remainder-to,omitempty"`
-
-	// \[rcv\] receiver's address.
-	Receiver string `json:"receiver"`
-}
-
-// TransactionSignature defines model for TransactionSignature.
-type TransactionSignature struct {
-
-	// \[lsig\] Programatic transaction signature.
-	//
-	// Definition:
-	// data/transactions/logicsig.go
-	Logicsig TransactionSignatureLogicsig `json:"logicsig,omitempty"`
-
-	// \[msig\] structure holding multiple subsignatures.
-	//
-	// Definition:
-	// crypto/multisig.go : MultisigSig
-	Multisig TransactionSignatureMultisig `json:"multisig,omitempty"`
-
-	// \[sig\] Standard ed25519 signature.
-	Sig []byte `json:"sig,omitempty"`
-}
-
-// TransactionSignatureLogicsig defines model for TransactionSignatureLogicsig.
-type TransactionSignatureLogicsig struct {
-
-	// \[arg\] Logic arguments, base64 encoded.
-	Args []string `json:"args,omitempty"`
-
-	// \[l\] Program signed by a signature or multi signature, or hashed to be the address of ana ccount. Base64 encoded TEAL program.
-	Logic []byte `json:"logic"`
-
-	// \[msig\] structure holding multiple subsignatures.
-	//
-	// Definition:
-	// crypto/multisig.go : MultisigSig
-	MultisigSignature TransactionSignatureMultisig `json:"multisig-signature,omitempty"`
-
-	// \[sig\] ed25519 signature.
-	Signature []byte `json:"signature,omitempty"`
-}
-
-// TransactionSignatureMultisig defines model for TransactionSignatureMultisig.
-type TransactionSignatureMultisig struct {
-
-	// \[subsig\] holds pairs of public key and signatures.
-	Subsignature []TransactionSignatureMultisigSubsignature `json:"subsignature,omitempty"`
-
-	// \[thr\]
-	Threshold uint64 `json:"threshold,omitempty"`
-
-	// \[v\]
-	Version uint64 `json:"version,omitempty"`
-}
-
-// TransactionSignatureMultisigSubsignature defines model for TransactionSignatureMultisigSubsignature.
-type TransactionSignatureMultisigSubsignature struct {
-
-	// \[pk\]
-	PublicKey []byte `json:"public-key,omitempty"`
-
-	// \[s\]
-	Signature []byte `json:"signature,omitempty"`
-}
-
-// Version defines model for Version.
-type Version struct {
-
-	// the current algod build version information.
-	Build       VersionBuild `json:"build"`
-	GenesisHash []byte       `json:"genesis-hash"`
-	GenesisId   string       `json:"genesis-id"`
-	Versions    []string     `json:"versions"`
 }
 
 // VersionBuild defines model for the current algod build version information.
@@ -590,89 +129,8 @@ type TxId []byte
 // TxType defines model for tx-type.
 type TxType string
 
-// AccountResponse defines model for AccountResponse.
-type AccountResponse struct {
-
-	// Account information at a given round.
-	//
-	// Definition:
-	// data/basics/userBalance.go : AccountData
-	Account Account `json:"account"`
-
-	// Round at which the results were computed.
-	CurrentRound uint64 `json:"current-round"`
-
-	// Used for pagination, when making another request provide this token with the next parameter.
-	NextToken string `json:"next-token"`
-}
-
-// AccountsResponse defines model for AccountsResponse.
-type AccountsResponse struct {
-	Accounts []Account `json:"accounts"`
-
-	// Round at which the results were computed.
-	CurrentRound uint64 `json:"current-round"`
-
-	// Used for pagination, when making another request provide this token with the next parameter.
-	NextToken string `json:"next-token"`
-}
-
-// AssetBalancesResponse defines model for AssetBalancesResponse.
-type AssetBalancesResponse struct {
-
-	// A simplified version of AssetHolding
-	Balances []MiniAssetHolding `json:"balances"`
-
-	// Round at which the results were computed.
-	CurrentRound uint64 `json:"current-round"`
-
-	// Used for pagination, when making another request provide this token with the next parameter.
-	NextToken string `json:"next-token"`
-}
-
-// AssetResponse defines model for AssetResponse.
-type AssetResponse struct {
-
-	// Specifies both the unique identifier and the parameters for an asset
-	Asset Asset `json:"asset"`
-
-	// Round at which the results were computed.
-	CurrentRound uint64 `json:"current-round"`
-}
-
-// AssetsResponse defines model for AssetsResponse.
-type AssetsResponse struct {
-	Assets []Asset `json:"assets"`
-
-	// Round at which the results were computed.
-	CurrentRound uint64 `json:"current-round"`
-
-	// Used for pagination, when making another request provide this token with the next parameter.
-	NextToken string `json:"next-token"`
-}
-
-// BlockResponse defines model for BlockResponse.
-type BlockResponse Block
-
 // HealthCheckResponse defines model for HealthCheckResponse.
 type HealthCheckResponse HealthCheck
-
-// ErrorResponse defines model for ErrorResponse.
-type ErrorResponse struct {
-	Data    *map[string]interface{} `json:"data,omitempty"`
-	Message string                  `json:"message"`
-}
-
-// TransactionsResponse defines model for TransactionsResponse.
-type TransactionsResponse struct {
-
-	// Round at which the results were computed.
-	CurrentRound uint64        `json:"current-round"`
-	Transactions []Transaction `json:"transactions"`
-
-	// Used for pagination, when making another request provide this token with the next parameter.
-	NextToken string `json:"next-token"`
-}
 
 // GetBlock response is returned by Block
 type GetBlockResponse = struct {
