+++ client/v2/common/models/transaction_asset_freeze.go
@@ -1,0 +1,15 @@
+package models
+
+// TransactionAssetFreeze fields for an asset freeze transaction.
+// Definition:
+// data/transactions/asset.go : AssetFreezeTxnFields
+type TransactionAssetFreeze struct {
+	// Address (fadd) Address of the account whose asset is being frozen or thawed.
+	Address string `json:"address,omitempty"`
+
+	// AssetId (faid) ID of the asset being frozen or thawed.
+	AssetId uint64 `json:"asset-id,omitempty"`
+
+	// NewFreezeStatus (afrz) The new freeze status.
+	NewFreezeStatus bool `json:"new-freeze-status,omitempty"`
+}
