+++ client/v2/common/models/transaction_signature_multisig_subsignature.go
@@ -1,0 +1,10 @@
+package models
+
+// TransactionSignatureMultisigSubsignature defines a model for TransactionSignatureMultisigSubsignature.
+type TransactionSignatureMultisigSubsignature struct {
+	// PublicKey (pk)
+	PublicKey []byte `json:"public-key,omitempty"`
+
+	// Signature (s)
+	Signature []byte `json:"signature,omitempty"`
+}
